Build and deploy with CircleCI
I've created a new IAM user with permission to a single bucket and
entered the keys in the CircleCI interface.

Using s3deploy instead of aws-cli because it uses checksums to compare
the files instead of last modified time and doesn't force the whole site
to be re-copied.

I could cache the bin directory for faster and more efficient builds.